Fixes to gamma correction and tonemapping in postfx

+++ Engine/Source/Platform/DirectX_Shared/Constant_Buffer_Types.h
@@ -58,5 +58,5 @@ struct CB_PS_PostFx
 	float innerRadius;
 	float outerRadius;
 	float opacity;
-	float vnEnabled;
+	int vnEnabled;
 };
